requirements.txt
streamlit
dotenv
pandas
openai
google-generativeai